Settings: system/custom radios and backend sliders
Appearance uses exclusive radios for system vs custom colors, tab bar
color, playlist playing-song color, and background type, plus sliders
for blur and opacity. Backend buffer, ReplayGain, EBU, and fade values
use sliders; ALSA plugins and None/ReplayGain/EBU normalization are
radios.

+++ src/engine/gstengine.cpp
@@ -37,7 +37,7 @@ bool GstEngine::Init() {
     const std::string mode = settings.Value("rgmode");
     replaygain_mode_ = (mode == "track" || mode == "1") ? 1 : settings.IntValue("rgmode", 0);
   }
-  replaygain_preamp_ = settings.IntValue("rgpreamp", 0);
+  replaygain_preamp_ = settings.DoubleValue("rgpreamp", 0);
   Settings eq;
   eq.BeginGroup(EqualizerPersist::kSettingsGroup);
   if (eq.Contains(EqualizerPersist::kEnableStereoBalancer) || eq.Contains(EqualizerPersist::kStereoBalance)) {
